requirements.txt
fastapi>=0.121.0
uvicorn[standard]>=0.39.0
pytest>=8.4.2
httpx>=0.28.1
websockets>=15.0.1
cryptography>=46.0.0
strawberry-graphql[fastapi]>=0.284.0
kafka-python>=2.2.15
